Move upcast conversion helpers for selection to upcasthelpers.js.
