+++ tests/conversion/upcast-selection-converters.js
@@ -11,10 +11,10 @@ import createViewRoot from '../view/_utils/createroot';
 import Model from '../../src/model/model';
 
 import Mapper from '../../src/conversion/mapper';
-import { convertSelectionChange } from '../../src/conversion/upcast-selection-converters';
 
 import { setData as modelSetData, getData as modelGetData } from '../../src/dev-utils/model';
 import { setData as viewSetData } from '../../src/dev-utils/view';
+import { convertSelectionChange } from '../../src/conversion/upcasthelpers';
 
 describe( 'convertSelectionChange', () => {
 	let model, view, viewDocument, mapper, convertSelection, modelRoot, viewRoot;
